#218 [refactor] 지역 조회 Service 이동
