+++ src/main/java/com/moddy/server/service/auth/AuthService.java
@@ -8,9 +8,7 @@ import com.moddy.server.common.util.SmsUtil;
 import com.moddy.server.common.util.VerificationCodeGenerator;
 import com.moddy.server.config.jwt.JwtService;
 import com.moddy.server.controller.auth.dto.response.LoginResponseDto;
-import com.moddy.server.controller.auth.dto.response.RegionResponse;
 import com.moddy.server.controller.designer.dto.response.UserCreateResponse;
-import com.moddy.server.domain.region.repository.RegionJpaRepository;
 import com.moddy.server.domain.user.User;
 import com.moddy.server.domain.user.repository.UserRepository;
 import com.moddy.server.external.kakao.service.KakaoSocialService;
@@ -19,9 +17,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
-import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 import static com.moddy.server.common.exception.enums.ErrorCode.INVALID_PHONE_NUMBER_EXCEPTION;
 import static com.moddy.server.common.exception.enums.ErrorCode.NOT_FOUND_VERIFICATION_CODE_EXCEPTION;
@@ -36,7 +32,6 @@ public class AuthService {
     private final JwtService jwtService;
     private final KakaoSocialService kakaoSocialService;
     private final UserRepository userRepository;
-    private final RegionJpaRepository regionJpaRepository;
 
     public LoginResponseDto login(final String baseUrl, final String kakaoCode) {
         String kakaoId = kakaoSocialService.getIdFromKakao(baseUrl, kakaoCode);
@@ -55,19 +50,6 @@ public class AuthService {
         }
         TokenPair tokenPair = jwtService.generateTokenPair(String.valueOf(user.get().getId()));
         return new LoginResponseDto(tokenPair.accessToken(), tokenPair.refreshToken(), user.get().getRole().name());
-    }
-
-    public List<RegionResponse> getRegionList() {
-
-        List<RegionResponse> regionResponseList = regionJpaRepository.findAll().stream().map(region -> {
-            RegionResponse regionResponse = new RegionResponse(
-                    region.getId(),
-                    region.getName()
-            );
-            return regionResponse;
-        }).collect(Collectors.toList());
-
-        return regionResponseList;
     }
 
     public UserCreateResponse createUserToken(String useId) {
